+++ v3/crates/sql/src/catalog/subgraph.rs
@@ -35,7 +35,7 @@ impl datafusion::SchemaProvider for catalog::model::WithSession<Subgraph> {
         name: &str,
     ) -> datafusion::Result<Option<Arc<dyn datafusion::TableProvider>>> {
         if let Some(model) = self.value.tables.get(name) {
-            let table = model::Table::new_no_args(self.value.metadata.clone(), model.clone());
+            let table = model::Table::new_no_args(self.value.metadata.clone(), model.clone())?;
             Ok(Some(Arc::new(table) as Arc<dyn datafusion::TableProvider>))
         } else {
             Ok(None)